requirements.txt
torch>=1.9.0
numpy
opencv-python
pillow
matplotlib
pytesseract
lgpio
smbus
Flask
email-validator
Werkzeug
itsdangerous
click
blinker
picamera2
adafruit-circuitpython-ssd1306
adafruit-circuitpython-framebuf
RPi.GPIO
